Fix long-range vdW correction for real systems.

+++ openmmapi/src/NonbondedForceImpl.cpp
@@ -259,7 +259,7 @@ double NonbondedForceImpl::calcDispersionCorrection(const System& system, const 
             if (useSwitch)
                 sum3 += count*epsilon*(evalIntegral(cutoff, switchDist, cutoff, sigma)-evalIntegral(switchDist, switchDist, cutoff, sigma));
         }
-    int numParticles = system.getNumParticles();
+    double numParticles = (double) system.getNumParticles();
     int numInteractions = (numParticles*(numParticles+1))/2;
     sum1 /= numInteractions;
     sum2 /= numInteractions;
